[#16] refactor: jwt 관련 코드 core 모듈로 이관
