+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/config/PropertiesConfig.java
@@ -1,8 +1,8 @@
-package org.nexters.jaknaesoserver.config;
+package org.nexters.jaknaesocore.config;
 
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ConfigurationPropertiesScan(basePackages = "org.nexters.jaknaesoserver")
+@ConfigurationPropertiesScan(basePackages = "org.nexters.jaknaesocore")
 public class PropertiesConfig {}
